backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
torch
sentence-transformers
chromadb
sqlalchemy
psycopg[binary]
transformers
datasets
peft
accelerate
textstat
pyspellchecker
